fixed a bug in zipfile handling
fixed various warnings


git-svn-id: http://developer.marklogic.com/svn/recordloader/trunk@498 e04f4502-82db-0310-b1af-f799f365da79

+++ src/java/com/marklogic/ps/Utilities.java
@@ -540,7 +540,7 @@ public class Utilities {
         byte[] bytes = _string.getBytes(_encoding);
         for (int i = 0; i < bytes.length; i++) {
             // bytes are signed: we want unsigned values
-            sb.append(Integer.toHexString((int) bytes[i] & 0xff));
+            sb.append(Integer.toHexString(bytes[i] & 0xff));
             if (i < bytes.length - 1) {
                 sb.append(" ");
             }
